+++ lib/loco_motion.rb
@@ -16,6 +16,7 @@ require "loco_motion/helpers"
 
 # Load patches
 require "loco_motion/patches/view_component/slotable_default_patch"
+require "loco_motion/patches/view_component/slot_loco_parent_patch"
 
 require "hero"
 require "daisy"